fix: change `uint128` -> `uint128`

+++ contracts/OrderBook.sol
@@ -472,11 +472,11 @@ abstract contract OrderBook is CloberOrderBook, ReentrancyGuard, RevertOnDelegat
 
     function indexToPrice(uint16 priceIndex) public view virtual returns (uint256);
 
-    function priceToIndex(uint128 price, bool roundingUp)
+    function priceToIndex(uint256 price, bool roundingUp)
         public
         view
         virtual
-        returns (uint16 index, uint128 correctedPrice);
+        returns (uint16 index, uint256 correctedPrice);
 
     function _cleanHeap(bool isBid) private {
         OctopusHeap.Core storage heap = _getHeap(isBid);
